fix syntax

+++ ft_substr.c
@@ -14,12 +14,12 @@
 ** External functs.
 ** malloc
 */
-//#include "libft.h"
+#include "libft.h"
 
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
 	char	*arr;
-	int		index;
+	size_t		index;
 
 	index = 0;
 	arr = (char *)malloc((len + 1) * sizeof(char));
